backend/requirements.txt
anthropic>=0.42.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
httpx>=0.27.0
pypdf>=5.0.0
